[TL;DR] Made find command smarter to reply with proper response

+++ src/services/find-query.ts
@@ -1,6 +1,7 @@
-import { Message } from "discord.js";
+import Discord, { Message } from "discord.js";
 import { injectable } from "inversify";
-import { FindHelper } from "../utils/find-helper";
+import { isString } from "../utils/extensions";
+import { FindHelper, QuestionResult } from "../utils/find-helper";
 import { StringUtils } from "../utils/string-utils";
 
 @injectable()
@@ -15,18 +16,37 @@ export class FindQuery {
     }
 
     public async command(message: Message, data: string): Promise<any> {
-        await message.react('👍')
         const msg = await message.channel.send("Searching 1/2")
-        const URL = await FindHelper.find(data);
-        await msg.edit("Searching 2/2")
-        const result = await FindHelper.captureJson(URL)
-        await msg.delete()
+        const findResult = await FindHelper.find(data);
+        if (findResult != null) {
+            await message.react('👍')
+            if (isString(findResult)) { // A perfect match has been found.
+                await msg.edit("Searching 2/2")
+                const result = await FindHelper.captureJson(findResult as string)
+                await msg.delete()
 
-        // Discord has limit of sending message length fewer than 2000
-        let answer = result.answer.replace(/&#39;/g, "'")
-        if (answer.length > 1300) {
-            answer = `${result.answer.substring(0, 1300)}\n\n${URL}`
+                // Discord has limit of sending message length fewer than 2000.
+                let answer = result.answer.replace(/&#39;/g, "'")
+                if (answer.length > 1300) {
+                    answer = `${result.answer.substring(0, 1300)}..., [read more](${result})`
+                }
+                return message.reply(`A perfect search has been found,\n\n**Q. ${result.question}**\n${answer}\n\n_- ${result.author}_`)
+            } else {
+                await msg.delete()
+                const array = findResult as QuestionResult[]
+                let answer = "Some similar searches has been found,\n\n"
+
+                // This will also handle Discord message character send limit.
+                for (let i = 0; i < array.length; i++) {
+                    if (answer.length > 1300) break;
+                    const r = array[i]
+                    answer += `**Q. ${r.question}**\n${r.shortAnswer} [read more](${r.uri})\n\n`
+                }
+                answer += "_Note: This searches might not be perfect, I'm just a bot not an AI :)_"
+                return message.reply(answer)
+            }
         }
-        return message.reply(`\n**Q. ${result.question}**\n${answer}\n\n_- ${result.author}_`)
+        await message.react('👎')
+        return msg.edit(`Question: **${data}** was not found`)
     }
 }
